backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pandas
scikit-learn
joblib
numpy
python-dotenv
